api_stridetastic/requirements.txt
python-dotenv
django
django-cors-headers
django-ninja
django-ninja-jwt
django-ninja-extra
djangorestframework
django-timescaledb
email-validator
django-unfold
celery
redis
django-celery-beat
django-redis
psycopg2
factory_boy
paho-mqtt
meshtastic
cryptography
pytest
pytest-django
pytest-cov
pytest-mock